Move actions and reducer to components directory
