+++ app/javascript/src/admin/components/clustering/actions.js
