+++ app/javascript/src/admin/components/clustering/reducer.js
